refactor(db): split config repository by domain
